Splitted NavigatorSymbols.
