+++ ts/masters/modules/gantt.src.ts
@@ -16,10 +16,11 @@ import '../../Series/Gantt/GanttSeries.js';
 import GanttChart from '../../Core/Chart/GanttChart.js';
 import Scrollbar from '../../Core/Scrollbar.js';
 import '../../Extensions/RangeSelector.js';
-import '../../Core/Navigator.js';
+import Navigator from '../../Core/Navigator.js';
 const G: AnyRecord = Highcharts;
 // Classes
 G.Scrollbar = Scrollbar;
+G.Navigator = Navigator;
 G.GanttChart = GanttChart;
 G.ganttChart = GanttChart.ganttChart;
 // Compositions
